backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic>=2.4.2,<3.0.0
pydantic-settings>=2.0.3,<3.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
redis==5.0.1
tenacity==8.2.3
httpx==0.25.1
aiosqlite==0.19.0
alembic==1.12.1
email-validator==2.1.0.post1
python-dotenv==1.0.0 